docs: add badges to 'README.md'
